// src/app/interfaces/employer-profile.interface.ts
export interface EmployerProfileInterface {
    fullname: string;
    photo_url: string;
    name: string;
    since: number;
    teamsize: string;
    uid: string;
    categories: string;
    description: string;
    userId: string;
    phone_number: string;
    email: string;
    website: string;
    country: string;
    city: string;
    address: string;
    lat: string;
    lon: string;
}